requirements.txt
accelerate==1.10.1
diffusers==0.27.2
easyocr==1.7.2
fastapi==0.108.0
huggingface-hub==0.25.2
numpy==1.26.4
opencv-python==4.11.0.86
Pillow==9.5.0
pydantic==2.12.5
python-multipart==0.0.20
safetensors==0.7.0
torch==2.8.0
torchvision==0.23.0
transformers==4.48.3
uvicorn==0.39.0
